design: 미션 현황 페이지 디자인 고도화
