+++ frontend/src/pages/MyMissionPage.tsx
@@ -1,6 +1,6 @@
 import MyMissionInProgress from '@/components/MyMission/MyMissionInProgress/MyMissionInProgress';
-import MyMissionCompleted from '@/components/MyMission/MyMissionCompleted/MyMissionCompleted';
 import * as S from './MyMissionPage.styled';
+import MyMissionCompleted from '@/components/MyMission/MyMissionCompleted/MyMissionCompleted';
 
 export default function MyMissionPage() {
   return (
